old/units_py/requirements.txt
attrs==18.1.0
mypy==0.600
typed-ast==1.1.0
